web: split orgs, maintainers, scans handlers out of server.go
server.go was 2101 lines and 43 handler methods. Move the three
cleanest domain clusters into their own files:

  orgs.go        orgsList, orgShow, plus parseAnyTime/sortSlice/orgRow
                 (used only here)
  maintainers.go maintainersList, maintainerShow, maintainerDoNotContact
  scans.go       jobs, scanShow, scanRetry, scansRetryFailed, scanCancel,
                 scanLog, scanSkillNames, retryFailedToast

Pure moves within the same package; no behaviour or signature change.
server.go is now 1568 lines. Repo and finding handlers stay where they
are for now since they share more local helpers; a follow-up can take
those once #208 settles.

docs/development.md project-layout table updated for the new files.
