+++ internal/web/server.go
@@ -485,322 +485,6 @@ func (s *Server) repoList(w http.ResponseWriter, r *http.Request) {
 	}
 }
 
-// parseAnyTime tries the handful of timestamp shapes SQLite gives us
-// back when a column is read via a raw SELECT (no type hint). Returns
-// (zero, false) for unparseable input so callers can decide what to do.
-func parseAnyTime(s string) (time.Time, bool) {
-	s = strings.TrimSpace(s)
-	if s == "" {
-		return time.Time{}, false
-	}
-	layouts := []string{
-		time.RFC3339Nano,
-		time.RFC3339,
-		"2006-01-02 15:04:05.999999999-07:00",
-		"2006-01-02 15:04:05.999999999Z07:00",
-		"2006-01-02 15:04:05-07:00",
-		"2006-01-02 15:04:05",
-	}
-	for _, layout := range layouts {
-		if t, err := time.Parse(layout, s); err == nil {
-			return t, true
-		}
-	}
-	return time.Time{}, false
-}
-
-// orgRow is one row on the orgs index — an aggregate over all repos
-// sharing the same Owner value.
-type orgRow struct {
-	Owner         string
-	Repos         int
-	FindingsTotal int
-	LastActivity  *time.Time
-}
-
-func (s *Server) orgsList(w http.ResponseWriter, r *http.Request) {
-	search := strings.TrimSpace(r.URL.Query().Get("q"))
-
-	// SQLite returns MAX() over a datetime column as a string; scan into
-	// a string and parse to *time.Time ourselves rather than fight GORM.
-	type aggRow struct {
-		Owner        string
-		Repos        int
-		LastActivity string
-	}
-	var aggs []aggRow
-	q := s.DB.Model(&db.Repository{}).
-		Select("owner, COUNT(*) AS repos, MAX(updated_at) AS last_activity").
-		Where("owner != ''").
-		Group("owner")
-	if search != "" {
-		q = q.Where("owner LIKE ?", "%"+search+"%")
-	}
-	q.Scan(&aggs)
-
-	// One grouped query gets finding totals per owner.
-	findingCounts := map[string]int{}
-	if len(aggs) > 0 {
-		type c struct {
-			Owner string
-			N     int
-		}
-		var counts []c
-		// LEFT JOIN scans so the COUNT only includes findings whose scan is
-		// the curated audit. Tool-scanner output (zizmor, semgrep) is shown
-		// per-repo in the Scanners tab, not in cross-org totals.
-		s.DB.Raw(`
-			SELECT r.owner, COUNT(f.id) AS n
-			FROM repositories r
-			LEFT JOIN findings f ON f.repository_id = r.id
-			LEFT JOIN scans s ON s.id = f.scan_id
-			WHERE r.owner != ''
-			  AND (s.skill_name IS NULL OR s.skill_name = '' OR s.skill_name = ?)
-			GROUP BY r.owner
-		`, deepDiveSkillName).Scan(&counts)
-		for _, x := range counts {
-			findingCounts[x.Owner] = x.N
-		}
-	}
-
-	rows := make([]orgRow, 0, len(aggs))
-	for _, a := range aggs {
-		row := orgRow{
-			Owner:         a.Owner,
-			Repos:         a.Repos,
-			FindingsTotal: findingCounts[a.Owner],
-		}
-		if t, ok := parseAnyTime(a.LastActivity); ok {
-			row.LastActivity = &t
-		}
-		rows = append(rows, row)
-	}
-
-	const nameSort = "name"
-	sort := r.URL.Query().Get("sort")
-	switch sort {
-	case "findings":
-		sortSlice(rows, func(a, b orgRow) bool { return a.FindingsTotal > b.FindingsTotal })
-	case "repos":
-		sortSlice(rows, func(a, b orgRow) bool { return a.Repos > b.Repos })
-	case defaultSort:
-		sortSlice(rows, func(a, b orgRow) bool {
-			if a.LastActivity == nil {
-				return false
-			}
-			if b.LastActivity == nil {
-				return true
-			}
-			return a.LastActivity.After(*b.LastActivity)
-		})
-	default:
-		sort = nameSort
-		sortSlice(rows, func(a, b orgRow) bool {
-			return strings.ToLower(a.Owner) < strings.ToLower(b.Owner)
-		})
-	}
-
-	s.render(w, r, "orgs.html", map[string]any{
-		"Orgs": rows, "Q": search, "Sort": sort,
-	})
-}
-
-// sortSlice is a tiny wrapper so the handler reads like `sortSlice(rows,
-// less)` without pulling sort.Slice's (i, j int) idiom into each case.
-func sortSlice[T any](s []T, less func(a, b T) bool) {
-	sort.Slice(s, func(i, j int) bool { return less(s[i], s[j]) })
-}
-
-func (s *Server) orgShow(w http.ResponseWriter, r *http.Request) {
-	owner := r.PathValue("login")
-	if owner == "" {
-		http.NotFound(w, r)
-		return
-	}
-
-	var repos []db.Repository
-	s.DB.Where("owner = ?", owner).Order("name").Find(&repos)
-	if len(repos) == 0 {
-		http.NotFound(w, r)
-		return
-	}
-	repoIDs := make([]uint, 0, len(repos))
-	for _, r := range repos {
-		repoIDs = append(repoIDs, r.ID)
-	}
-
-	// Per-repo finding count for the repos table, plus the full finding
-	// list for the Findings tab. Scanner skills (zizmor, semgrep) are kept
-	// out of both so the org summary matches the repo Findings tab; the
-	// per-repo Scanners tab is where that output lives.
-	findingCounts := map[uint]int{}
-	type rowCount struct {
-		RepositoryID uint
-		N            int
-	}
-	var counts []rowCount
-	s.DB.Model(&db.Finding{}).
-		Select("repository_id, COUNT(*) AS n").
-		Where("repository_id IN ?", repoIDs).
-		Where("scan_id IN (?)", deepDiveScanIDs(s.DB)).
-		Group("repository_id").Scan(&counts)
-	for _, c := range counts {
-		findingCounts[c.RepositoryID] = c.N
-	}
-
-	const orgTabLimit = 200
-	// Sort by severity (Critical→High→Medium→Low), then newest first
-	// within a severity. Purely alphabetical severity would put Low
-	// before Medium, which misreads for a stakeholder scanning the tab.
-	var findings []db.Finding
-	s.DB.Where("repository_id IN ?", repoIDs).
-		Where("scan_id IN (?)", deepDiveScanIDs(s.DB)).
-		Order(severityOrder).Order("id desc").
-		Limit(orgTabLimit).Find(&findings)
-	reposByID := loadRepoMap(s.DB, findings, findingRepoID)
-
-	var advisories []db.Advisory
-	s.DB.Where("repository_id IN ?", repoIDs).Order("cvss_score desc").
-		Limit(orgTabLimit).Find(&advisories)
-	advisoryRepos := loadRepoMap(s.DB, advisories, advisoryRepoID)
-
-	var maintainers []db.Maintainer
-	s.DB.Joins("JOIN repository_maintainers rm ON rm.maintainer_id = maintainers.id").
-		Where("rm.repository_id IN ?", repoIDs).
-		Distinct().Order("maintainers.name").Find(&maintainers)
-
-	s.render(w, r, "org_show.html", map[string]any{
-		"Owner":         owner,
-		"Repos":         repos,
-		"FindingCounts": findingCounts,
-		"Findings":      findings,
-		"FindingRepos":  reposByID,
-		"Advisories":    advisories,
-		"AdvisoryRepos": advisoryRepos,
-		"Maintainers":   maintainers,
-	})
-}
-
-func (s *Server) maintainersList(w http.ResponseWriter, r *http.Request) {
-	q := s.DB.Model(&db.Maintainer{})
-	status := r.URL.Query().Get("status")
-	if status != "" {
-		q = q.Where("status = ?", status)
-	}
-	search := strings.TrimSpace(r.URL.Query().Get("q"))
-	if search != "" {
-		like := "%" + search + "%"
-		q = q.Where("login LIKE ? OR name LIKE ? OR email LIKE ? OR company LIKE ? OR notes LIKE ?",
-			like, like, like, like, like)
-	}
-
-	const nameSort = "name"
-	sort := r.URL.Query().Get("sort")
-	switch sort {
-	case "login":
-		q = q.Order("login")
-	case "status":
-		q = q.Order("status, name")
-	case "newest":
-		q = q.Order("id desc")
-	default:
-		sort = nameSort
-		// Push empty names to the end instead of the front.
-		q = q.Order("CASE WHEN name = '' THEN 1 ELSE 0 END, name, login")
-	}
-
-	var total int64
-	q.Count(&total)
-	page := paginate(r, total)
-
-	var rows []db.Maintainer
-	q.Preload("Repositories").
-		Limit(perPage).Offset((page.N - 1) * perPage).Find(&rows)
-
-	// Batch-count findings across each maintainer's linked repositories
-	// in a single grouped query rather than one query per maintainer.
-	findingCounts := map[uint]int{}
-	if len(rows) > 0 {
-		ids := make([]uint, 0, len(rows))
-		for _, m := range rows {
-			ids = append(ids, m.ID)
-		}
-		type row struct {
-			MaintainerID uint
-			N            int
-		}
-		var counts []row
-		// LEFT JOIN scans so the COUNT only includes deep-dive findings.
-		// Scanner output (zizmor, semgrep) is per-repo lint noise and
-		// shouldn't drive maintainer routing.
-		s.DB.Raw(`
-			SELECT rm.maintainer_id, COUNT(f.id) AS n
-			FROM repository_maintainers rm
-			LEFT JOIN findings f ON f.repository_id = rm.repository_id
-			LEFT JOIN scans s ON s.id = f.scan_id
-			WHERE rm.maintainer_id IN ?
-			  AND (s.skill_name IS NULL OR s.skill_name = '' OR s.skill_name = ?)
-			GROUP BY rm.maintainer_id
-		`, ids, deepDiveSkillName).Scan(&counts)
-		for _, c := range counts {
-			findingCounts[c.MaintainerID] = c.N
-		}
-	}
-
-	s.render(w, r, "maintainers.html", map[string]any{
-		"Maintainers":   rows,
-		"Page":          page,
-		"Status":        status,
-		"Q":             search,
-		"Sort":          sort,
-		"FindingCounts": findingCounts,
-	})
-}
-
-// maintainerDoNotContact flips the DoNotContact flag on a maintainer.
-// Toggle semantics — form posts an explicit `value` of "true" or "false".
-func (s *Server) maintainerDoNotContact(w http.ResponseWriter, r *http.Request) {
-	id, _ := strconv.Atoi(r.PathValue("id"))
-	var m db.Maintainer
-	if err := s.DB.First(&m, id).Error; err != nil {
-		http.NotFound(w, r)
-		return
-	}
-	value := r.FormValue("value") == "true"
-	if err := s.DB.Model(&db.Maintainer{}).Where("id = ?", m.ID).
-		Update("do_not_contact", value).Error; err != nil {
-		http.Error(w, err.Error(), http.StatusInternalServerError)
-		return
-	}
-	s.redirect(w, r, fmt.Sprintf("/maintainers/%d", m.ID))
-}
-
-func (s *Server) maintainerShow(w http.ResponseWriter, r *http.Request) {
-	var m db.Maintainer
-	if err := s.DB.Preload("Repositories").First(&m, r.PathValue("id")).Error; err != nil {
-		http.NotFound(w, r)
-		return
-	}
-	// Gather findings across all their repos
-	repoIDs := make([]uint, 0, len(m.Repositories))
-	for _, repo := range m.Repositories {
-		repoIDs = append(repoIDs, repo.ID)
-	}
-	var findings []db.Finding
-	if len(repoIDs) > 0 {
-		// Same filter the Findings tab applies elsewhere: deep-dive only,
-		// keeping scanner noise off the maintainer view used for disclosure
-		// routing.
-		s.DB.Where("repository_id IN ?", repoIDs).
-			Where("scan_id IN (?)", deepDiveScanIDs(s.DB)).
-			Order("id desc").Find(&findings)
-	}
-	reposByID := loadRepoMap(s.DB, findings, findingRepoID)
-	s.render(w, r, "maintainer_show.html", map[string]any{
-		"M": m, "Findings": findings, "Repos": reposByID,
-	})
-}
-
 // loadRepoMap batch-loads the repositories referenced by a slice of
 // rows and returns a map keyed by repository ID. The repoID accessor
 // abstracts over Finding.RepositoryID, Advisory.RepositoryID and any
@@ -1326,70 +1010,6 @@ func (s *Server) findingShow(w http.ResponseWriter, r *http.Request) {
 	s.render(w, r, "finding_show.html", data)
 }
 
-func (s *Server) jobs(w http.ResponseWriter, r *http.Request) {
-	q := s.DB.Model(&db.Scan{})
-	skillName := r.URL.Query().Get("skill")
-	if skillName != "" {
-		q = q.Where("skill_name = ?", skillName)
-	}
-	status := r.URL.Query().Get("status")
-	if status != "" {
-		q = q.Where("status = ?", status)
-	}
-
-	sort := r.URL.Query().Get("sort")
-	switch sort {
-	case "skill":
-		q = q.Order("skill_name, id desc")
-	case "status":
-		q = q.Order("status, id desc")
-	case sortRepository:
-		q = q.Joins("Repository").Order("`Repository`.name, scans.id desc")
-	default:
-		sort = defaultSort
-		q = q.Order("status_priority, scans.id desc")
-	}
-
-	var total int64
-	q.Count(&total)
-	page := paginate(r, total)
-
-	var scans []db.Scan
-	q.Preload("Repository").
-		Limit(perPage).Offset((page.N - 1) * perPage).Find(&scans)
-
-	skillNames := s.scanSkillNames()
-
-	anySubPath := false
-	for _, sc := range scans {
-		if sc.SubPath != "" {
-			anySubPath = true
-			break
-		}
-	}
-	s.render(w, r, "jobs.html", map[string]any{
-		"Scans": scans, "Page": page,
-		"Skill": skillName, "Status": status, "Sort": sort, "Skills": skillNames,
-		"AnySubPath": anySubPath,
-	})
-}
-
-const skillNamesCacheTTL = 30 * time.Second
-
-func (s *Server) scanSkillNames() []string {
-	s.skillNamesMu.Lock()
-	defer s.skillNamesMu.Unlock()
-	if time.Now().Before(s.skillNamesTTL) {
-		return s.skillNamesCache
-	}
-	var names []string
-	s.DB.Model(&db.Scan{}).Where("skill_name != ''").Distinct("skill_name").
-		Order("skill_name").Pluck("skill_name", &names)
-	s.skillNamesCache = names
-	s.skillNamesTTL = time.Now().Add(skillNamesCacheTTL)
-	return names
-}
-
 func (s *Server) repoCreate(w http.ResponseWriter, r *http.Request) {
 	input, err := ParseRepoInput(r.FormValue("url"))
 	if err != nil {
@@ -1691,159 +1311,6 @@ func (s *Server) repoDisclosureChannel(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 	s.redirect(w, r, fmt.Sprintf("/repositories/%d", repo.ID))
-}
-
-func (s *Server) scanShow(w http.ResponseWriter, r *http.Request) {
-	var scan db.Scan
-	if err := s.DB.Preload("Repository").Preload("Findings").First(&scan, r.PathValue("id")).Error; err != nil {
-		http.NotFound(w, r)
-		return
-	}
-	s.render(w, r, "scan_show.html", map[string]any{"Scan": scan})
-}
-
-func (s *Server) scanRetry(w http.ResponseWriter, r *http.Request) {
-	scan, ok := loadByID[db.Scan](s, w, r)
-	if !ok {
-		return
-	}
-	if scan.Kind != worker.JobSkill || scan.SkillID == nil {
-		http.Error(w, "scan cannot be retried: no skill reference", http.StatusBadRequest)
-		return
-	}
-	newID, err := s.enqueueSkillWith(r.Context(), scan.RepositoryID, *scan.SkillID, ScanOpts{
-		Model:     scan.Model,
-		FindingID: scan.FindingID,
-		SubPath:   scan.SubPath,
-		Ref:       scan.Ref,
-	})
-	if err != nil {
-		http.Error(w, err.Error(), http.StatusInternalServerError)
-		return
-	}
-	s.redirect(w, r, fmt.Sprintf("/scans/%d", newID))
-}
-
-func (s *Server) scansRetryFailed(w http.ResponseWriter, r *http.Request) {
-	skillName := r.URL.Query().Get("skill")
-	repoID, _ := strconv.Atoi(r.URL.Query().Get("repository"))
-	q := s.DB.Model(&db.Scan{}).
-		Where("status = ? AND kind = ? AND skill_id IS NOT NULL", db.ScanFailed, worker.JobSkill)
-	if skillName != "" {
-		q = q.Where("skill_name = ?", skillName)
-	}
-	if repoID > 0 {
-		q = q.Where("repository_id = ?", repoID)
-	}
-
-	var totalFailed int64
-	q.Count(&totalFailed)
-
-	// Skip any failed scan that has a later scan with the same
-	// (repository, skill, sub_path, ref, finding_id) tuple already in
-	// queued/running/done.
-	var scans []db.Scan
-	err := q.Select("id, repository_id, skill_id, model, finding_id, sub_path, ref").
-		Where(`NOT EXISTS (
-			SELECT 1 FROM scans n
-			WHERE n.id > scans.id
-			  AND n.repository_id = scans.repository_id
-			  AND COALESCE(n.skill_id, 0) = COALESCE(scans.skill_id, 0)
-			  AND COALESCE(n.sub_path, '') = COALESCE(scans.sub_path, '')
-			  AND COALESCE(n.ref, '') = COALESCE(scans.ref, '')
-			  AND COALESCE(n.finding_id, 0) = COALESCE(scans.finding_id, 0)
-			  AND n.status IN ?
-		)`, []db.ScanStatus{db.ScanQueued, db.ScanRunning, db.ScanDone}).
-		Find(&scans).Error
-	if err != nil {
-		http.Error(w, err.Error(), http.StatusInternalServerError)
-		return
-	}
-
-	var retried, errored int
-	for _, sc := range scans {
-		if _, err := s.enqueueSkillWith(r.Context(), sc.RepositoryID, *sc.SkillID, ScanOpts{
-			Model:     sc.Model,
-			FindingID: sc.FindingID,
-			SubPath:   sc.SubPath,
-			Ref:       sc.Ref,
-		}); err != nil {
-			errored++
-			continue
-		}
-		retried++
-	}
-	skipped := int(totalFailed) - retried - errored
-
-	setFlash(w, retryFailedToast(retried, skipped, errored))
-	// Repo-scoped retries return to that repo's Scans tab so the operator
-	// stays in context; otherwise we send them to the global jobs list
-	// filtered to failed.
-	target := "/scans?status=failed"
-	if repoID > 0 {
-		target = fmt.Sprintf("/repositories/%d#rt3", repoID)
-	} else if skillName != "" {
-		target += "&skill=" + url.QueryEscape(skillName)
-	}
-	s.redirect(w, r, target)
-}
-
-func retryFailedToast(retried, skipped, errored int) Flash {
-	if retried == 0 && skipped == 0 && errored == 0 {
-		return Flash{Category: "success", Title: "No failed scans to retry"}
-	}
-	parts := []string{fmt.Sprintf("%d retried", retried)}
-	if skipped > 0 {
-		parts = append(parts, fmt.Sprintf("%d skipped", skipped))
-	}
-	if errored > 0 {
-		parts = append(parts, fmt.Sprintf("%d errored", errored))
-	}
-	cat := "success"
-	switch {
-	case errored > 0:
-		cat = "error"
-	case retried == 0:
-		cat = "warning"
-	}
-	return Flash{Category: cat, Title: strings.Join(parts, ", ")}
-}
-
-func (s *Server) scanCancel(w http.ResponseWriter, r *http.Request) {
-	scan, ok := loadByID[db.Scan](s, w, r)
-	if !ok {
-		return
-	}
-	if scan.Status.Terminal() {
-		http.Error(w, "scan already finished", http.StatusBadRequest)
-		return
-	}
-	if !s.Worker.Cancel(scan.ID) {
-		// Not in flight: mark the row so the queue handler drops it on pickup.
-		s.DB.Model(&scan).Updates(map[string]any{
-			"status":      db.ScanCancelled,
-			"error":       "cancelled by user",
-			"finished_at": new(time.Now()),
-		})
-	}
-	s.redirect(w, r, fmt.Sprintf("/scans/%d", scan.ID))
-}
-
-// scanLog returns just the <pre> log block. The scan page polls this with
-// hx-trigger while the scan is running so the operator can watch claude work.
-func (s *Server) scanLog(w http.ResponseWriter, r *http.Request) {
-	scan, ok := loadByID[db.Scan](s, w, r)
-	if !ok {
-		return
-	}
-	if scan.Status != db.ScanQueued && scan.Status != db.ScanRunning {
-		// Tell htmx to do a full refresh so the report renders.
-		w.Header().Set("HX-Refresh", "true")
-	}
-	w.Header().Set("Content-Type", "text/plain; charset=utf-8")
-	if err := s.tmpl.ExecuteTemplate(w, "scan_log.html", scan); err != nil {
-		http.Error(w, err.Error(), http.StatusInternalServerError)
-	}
 }
 
 // ScanOpts carries the optional inputs to an enqueue call. Keeps the
